+++ example/android/app/src/main/kotlin/com/example/esim_installer_example/MainActivity.kt
@@ -1,5 +1,7 @@
-package com.example.esim_installer_example
+package com.esim.installer
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity : FlutterActivity()
+class MainActivity: FlutterActivity() {
+    // Keep this empty unless you need to add customizations
+}